Fixed Branch+Link bug in assembly

+++ slh-dsa/armv8/SHAKE256/KECCAK.h
@@ -6,4 +6,4 @@
 
 void KeccakF1600_StatePermute_ARMv8A(uint64_t *ptr);
 
-void Keccak_Inc_Absorb_ARMv8A(uint64_t *s_inc, const uint8_t *m, size_t mlen);
+void Keccak_Inc_Absorb_ARMv8A(uint64_t *s_inc, const uint8_t *m, uint64_t mlen);
